# Conflicts:
#	app/src/main/java/com/example/seniorproject/MainForum/Fragments/FragmentHome.kt
#	app/src/main/java/com/example/seniorproject/data/Firebase/FirebaseData.kt
#	app/src/main/java/com/example/seniorproject/data/repositories/PostRepository.kt
#	app/src/main/java/com/example/seniorproject/viewModels/HomeFragmentViewModel.kt

+++ app/src/main/java/com/example/seniorproject/MainForum/Posts/ClickedPost.kt
@@ -1,7 +1,13 @@
 package com.example.seniorproject.MainForum.Posts
 
+import android.content.DialogInterface
+import android.graphics.Canvas
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.graphics.drawable.Drawable
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
@@ -20,6 +26,9 @@ import javax.inject.Inject
 class ClickedPost : AppCompatActivity() {
 
     private lateinit var adapter: CommentsAdapter
+    var swipeBackground: ColorDrawable = ColorDrawable(Color.parseColor("#FF0000"))
+    lateinit var deleteIcon: Drawable
+
     @Inject
     lateinit var factory: ViewModelProvider.Factory
     lateinit var myViewModel: ClickedPostViewModel
@@ -57,11 +66,12 @@ class ClickedPost : AppCompatActivity() {
         refreshView.setColorSchemeColors(ContextCompat.getColor(this, R.color.white))
 
         refreshView.setOnRefreshListener {
-            comment_RecyclerView.adapter = adapter
+            comment_RecyclerView.adapter = CommentsAdapter(this, myViewModel.getComments(), title, text, author, crn)
             refreshView.isRefreshing = false
         }
 
 
+        deleteIcon = ContextCompat.getDrawable(this, R.drawable.ic_delete_24px)!!
 
         val itemTouchHelperCallback = object: ItemTouchHelper.SimpleCallback(0,ItemTouchHelper.LEFT )
         {
@@ -70,11 +80,66 @@ class ClickedPost : AppCompatActivity() {
                 val commentkey: String? =adapter.removeItem(viewHolders as CommentsAdapter.CustomViewHolders, position)
               //  val commentkeyheader: String? =adapter.removeItemHead(viewHolders as , position)
 
+                var builder = AlertDialog.Builder(this@ClickedPost, R.style.AppTheme_AlertDialog)
 
+                //.getStringExtra("Classkey")
+                //val postkey = intent.getStringExtra("author")
+                //myViewModel.deletePost(postkey!!, className)
+                //myViewModel.deletePost()
+                builder.setTitle("Are you sure?")
+                builder.setMessage("You cannot restore comments that have been deleted.")
+                builder.setPositiveButton("DELETE",
+                    { dialogInterface: DialogInterface?, i: Int ->
+                        myViewModel.deleteComment(intent.getStringExtra("Classkey"), intent.getStringExtra("crn"),commentkey!!)
+                    })
+                builder.setNegativeButton("CANCEL",
+                    { dialogInterface: DialogInterface?, i: Int ->
+                        builder.setCancelable(true)
+                    })
+
+                val msgdialog: AlertDialog = builder.create()
+
+                msgdialog.show()
+
+                adapter.notifyItemRemoved(viewHolders.adapterPosition)
+                adapter.notifyItemRangeChanged(viewHolders.adapterPosition, adapter.itemCount)
+                comment_RecyclerView.adapter = adapter
                 //adapter.removeItem(viewHolders as CustomViewHolders)
-                myViewModel.deleteComment(intent.getStringExtra("Classkey"), intent.getStringExtra("crn"),commentkey!!)
+                //myViewModel.deleteComment(intent.getStringExtra("Classkey"), intent.getStringExtra("crn"),commentkey!!)
               //  myViewModel.deleteComment(intent.getStringExtra("Classkey"), intent.getStringExtra("crn"),commentkeyheader!!)
 
+            }
+
+            override fun onChildDraw(
+                c: Canvas,
+                recyclerView: RecyclerView,
+                viewHolder: RecyclerView.ViewHolder,
+                dX: Float,
+                dY: Float,
+                actionState: Int,
+                isCurrentlyActive: Boolean
+            ) {
+                val itemView = viewHolder.itemView
+                val iconMargin = (itemView.height - deleteIcon.intrinsicHeight)/2
+
+                if(dX < 0) {
+                    swipeBackground.setBounds(itemView.right + dX.toInt(), itemView.top, itemView.right, itemView.bottom)
+                    deleteIcon.setBounds(itemView.right - iconMargin - deleteIcon.intrinsicWidth, itemView.top + iconMargin,
+                        itemView.right - iconMargin, itemView.bottom - iconMargin)
+                }
+
+                swipeBackground.draw(c)
+                deleteIcon.draw(c)
+
+                super.onChildDraw(
+                    c,
+                    recyclerView,
+                    viewHolder,
+                    dX,
+                    dY,
+                    actionState,
+                    isCurrentlyActive
+                )
             }
 
             override fun onMove(
